Git manifest provider support for one repo multiple packages
Adds supports for packages like [ros-navigation](https://github.com/ros-planning/navigation) where there are multiple packages in one repo. In line with PR #126, there is a use case where the source repository can also be the release repository. This implies getting the manifest from a subfolder rather than from the root. Notice this is not a problem if the release repository (as designed by bloom) collects every source subfolder into a branch for that single package.

The PR sets the default behavior as of today and in case of failure will check in the subfolder. If not, will throw the exception as expected.